quality: Fix GCC 5.1 many warnings
 * Add 'final' for private classes
 * Add some missing override
 * Fix a constness issue
 * Quiet some float conversion warnings
 * Fix a scope issue in CMakeLists.txt
 
 This also fix a build problem with LLVM

Refs #74146

+++ src/media/avrecording.cpp
@@ -242,7 +242,7 @@ void Media::AVRecordingPrivate::notifySeek(int position, int size)
    //Update the metadata
    m_Duration = size/1000;
    m_Position = ((double)position)/((double)size);
-   m_Elapsed  = m_Duration * m_Position;
+   m_Elapsed  = m_Duration * static_cast<int>(m_Position);
    m_Left     = m_Duration - m_Elapsed ;
 
    if (oldDuration != m_Duration)
